+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - add support for relative paths to `--serve`
+- add `--resize`/`--no-resize` cli option to allow/prevent browser window resizing
+- add `--frame`/`--no-frame` cli option to show/hide browser window frame
+
+### Changed
+
+- decouple `--transparent` from window frame display and window resizing
 
 ## [0.14.1] 2020-12-02